# Palo Alto Networks NetSec-Pro – P
this is a 240 questions mock exam for simulation. Questions was generated by AI using the blueprint information and official Palo Alto Networks documentation. 
The main reason to have this is to help the student to test their knowledge based on the PALO ALTO Networks trainning track, official trainning and official documentation. 
Use under your own risk.